refactor(backend): refactor function getSqliteDatabasePath()
